app/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
djangorestframework==3.14.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
multidict==6.0.2
openpyxl==3.0.10
packaging==21.3
psycopg2==2.9.4
pydantic==1.10.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
